\name{is.cydquantmod}
\alias{is.cydquantmod}
\alias{is.cydquantmodResults}
\title{ Test If Object of Type cydquantmod }
\description{
Test if object is of type \code{cydquantmod} or
\code{cydquantmodResults}.
}
\usage{
is.cydquantmod(x)
is.cydquantmodResults(x)
}
\arguments{
  \item{x}{ object to test }
}
\value{
Boolean TRUE or FALSE
}
\author{ Jeffrey A. Ryan }
\seealso{ \code{\link{specifyModel}}, \code{\link{tradeModel}} }
\keyword{ utilities }
