\name{cydquantmod-package}
\alias{cydquantmod-package}
\alias{cydquantmod}
\alias{cydquantmodenv}
\docType{package}
\title{
Quantitative Financial Modelling Framework
}
\description{
Quantitative Financial Modelling and Trading Framework for R
}
\details{
\packageDESCRIPTION{cydquantmod}

The cydquantmod package for R is designed to
assist the quantitative trader in the
development, testing, and deployment
of statistically based trading models.


\emph{What cydquantmod IS}

A rapid prototyping environment, with comprehensive
tools for data management and visualization.
where quant traders can quickly
and cleanly explore and build trading models.


\emph{What cydquantmod is NOT}

A replacement for anything statistical.
It has no 'new' modelling routines or
analysis tool to speak of. It does now
offer charting not currently available
elsewhere in R, but most everything else
is more of a wrapper to what you already
know and love about the langauge
and packages you currently use.

cydquantmod makes modelling easier by removing
the repetitive workflow issues surrounding
data management, modelling interfaces,
and performance analysis. 

}
\author{
Jeffrey A. Ryan

Maintainer: Joshua M. Ulrich <josh.m.ulrich@gmail.com>
}
\keyword{ package }
